treewide: replace m.stop_propagation() with a @stop_here decorator
This way we can ensure that the message is stopped from propagating.
The current implementation may not be good for code with returns, as
the return would stop execution before propagation is stopped. An
alternative to this would be to call m.stop_propagation() every time
instead of a return, but this doesn't look like a clean way. Another
way would be to wrap all the code in a try...finally block, but this
is not a good option either.